Fix deleting nested tracks

+++ src/music/serializers.py
@@ -46,22 +46,24 @@ class AlbumSerializer(serializers.ModelSerializer):
         return album
 
     def update(self, instance, validated_data):
-        tracks_data = validated_data.pop('tracks', None)
-        super(self.__class__, self).update(instance, validated_data)
-        if tracks_data:
-            track_to_delete = instance.tracks.exclude(pk__in=[item.get('id', None) for item in tracks_data])
+        if 'tracks' in validated_data:
+            tracks_data = validated_data.pop('tracks', None)
+            if tracks_data:
+                track_to_delete = instance.tracks.exclude(pk__in=[item.get('id', None) for item in tracks_data])
 
-            for item in tracks_data:
-                item_id = item.get('id', None)
-                if item_id:
-                    track_item = Track.objects.get(id=item_id, album=instance)
-                    track_item.title = item.get('title', track_item.title)
-                    track_item.save()
-                else:
-                    Track.objects.create(album=instance, **item)
-        else:
-            track_to_delete = instance.tracks.all()
-        # Delete tracks which are not exists in validated_data
-        if track_to_delete:
-            track_to_delete.delete()
+                for item in tracks_data:
+                    item_id = item.get('id', None)
+                    if item_id:
+                        track_item = Track.objects.get(id=item_id, album=instance)
+                        track_item.title = item.get('title', track_item.title)
+                        track_item.save()
+                    else:
+                        Track.objects.create(album=instance, **item)
+            else:
+                track_to_delete = instance.tracks.all()
+            # Delete tracks which are not exists in validated_data
+            if track_to_delete:
+                track_to_delete.delete()
+
+        super(self.__class__, self).update(instance, validated_data)
         return instance
